requirements.txt
Flask==2.2.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
PyYAML==6.0.1
SQLAlchemy==2.0.19
gunicorn==21.2.0
python-gitlab==3.15.0
requests==2.28.2
spdx-tools==0.8.1
tmt==1.31.0
tmt[provision-container]
Werkzeug==2.2.2
